progress

+++ back/controllers/courseController.js
@@ -1,9 +1,12 @@
 const { CourseService } = require("../services");
 const { validationResult } = require("express-validator");
-const a =require("../providers/categoryProvider")
-
+const a = require("../providers/categoryProvider");
 
 const createCourse = async (req, res) => {
+  const result = validationResult(req);
+  if (!result.isEmpty()) {
+     return res.status(403).send({ errors: result.array() });
+  }
   const {
     name,
     description,
@@ -14,27 +17,35 @@ const createCourse = async (req, res) => {
     price,
     requirement,
     teacher,
-    CourseCategoryId
+    CourseCategoryName,
   } = req.body;
-  const newCourse = await CourseService.createCourse({
-    name,
-    description,
-    maxStudents,
-    start,
-    end,
-    active,
-    price,
-    requirement,
-    teacher,
-    CourseCategoryId
-  });
 
-  res.status(201).json(newCourse);
+  try {
+    const newCourse = await CourseService.createCourse({
+      name,
+      description,
+      maxStudents,
+      start,
+      end,
+      active,
+      price,
+      requirement,
+      teacher,
+      CourseCategoryName,
+    });
+
+    res.status(201).json(newCourse);
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
 };
 const getByIdCourse = async (req, res) => {
-
+  const result = validationResult(req);
+  if (!result.isEmpty()) {
+     return res.status(403).send({ errors: result.array() });
+  }
   const CourseId = req.params.CourseId;
-  console.log(CourseId)
+  console.log(CourseId);
   try {
     const Course = await CourseService.getCourse(CourseId);
     res.status(200).json(Course);
@@ -44,6 +55,10 @@ const getByIdCourse = async (req, res) => {
 };
 
 const getCourses = async (req, res) => {
+  const result = validationResult(req);
+  if (!result.isEmpty()) {
+     return res.status(403).send({ errors: result.array() });
+  }
   try {
     const Course = await CourseService.getCourses();
     res.status(200).json(Course);
@@ -53,10 +68,10 @@ const getCourses = async (req, res) => {
 };
 
 const updateCourse = async (req, res) => {
-  //const result = validationResult(req);
-  //if (!result.isEmpty()) {
- //   return res.status(403).send({ errors: result.array() });
-  //}
+  const result = validationResult(req);
+  if (!result.isEmpty()) {
+     return res.status(403).send({ errors: result.array() });
+  }
   const CourseId = req.params.CourseId;
   const {
     name,
@@ -67,8 +82,7 @@ const updateCourse = async (req, res) => {
     active,
     price,
     requirement,
-    teacher
-    //CourseCategory,
+    teacher,
   } = req.body;
   try {
     const newCourse = await CourseService.updateCourse(CourseId, {
@@ -80,8 +94,7 @@ const updateCourse = async (req, res) => {
       active,
       price,
       requirement,
-      teacher
-      //CourseCategory,
+      teacher,
     });
     res.status(200).json(newCourse);
   } catch (error) {
@@ -90,6 +103,10 @@ const updateCourse = async (req, res) => {
 };
 
 const deleteCourse = async (req, res) => {
+  const result = validationResult(req);
+  if (!result.isEmpty()) {
+     return res.status(403).send({ errors: result.array() });
+  }
   const CourseId = req.params.CourseId;
   try {
     const Course = await CourseService.deleteCourse(CourseId);
