+++ back/controllers/scheduleController.js
@@ -1,6 +1,11 @@
 const { ScheduleService } = require("../services");
+const { validationResult } = require("express-validator");
 
 const getSchedules = async (req, res) => {
+  const result = validationResult(req);
+  if (!result.isEmpty()) {
+     return res.status(403).send({ errors: result.array() });
+  }
   try {
     const Schedules = await ScheduleService.getSchedules();
     res.status(200).json(Schedules);
@@ -9,17 +14,30 @@ const getSchedules = async (req, res) => {
   }
 };
 const createSchedule = async (req, res) => {
-  const {start,end, active, where  } = req.body;
-  console.log(req.body);
+  const result = validationResult(req);
+  if (!result.isEmpty()) {
+     return res.status(403).send({ errors: result.array() });
+  }
+  const { start, end, active, where } = req.body;
+  try {
+    const newSchedule = await ScheduleService.createSchedule({
+      start,
+      end,
+      active,
+      where,
+    });
 
-  const newSchedule = await ScheduleService.createSchedule({
-    start,end, active, where 
-  });
-
-  res.status(201).json(newSchedule);
+    res.status(201).json(newSchedule);
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
 };
 
 const getByIdSchedule = async (req, res) => {
+  const result = validationResult(req);
+  if (!result.isEmpty()) {
+     return res.status(403).send({ errors: result.array() });
+  }
   const ScheduleId = req.params.ScheduleId;
   try {
     const Schedule = await ScheduleService.getSchedule(ScheduleId);
@@ -30,11 +48,18 @@ const getByIdSchedule = async (req, res) => {
 };
 
 const updateSchedule = async (req, res) => {
+  const result = validationResult(req);
+  if (!result.isEmpty()) {
+     return res.status(403).send({ errors: result.array() });
+  }
   const ScheduleId = req.params.ScheduleId;
-  const { start,end, active, where  } = req.body;
+  const { start, end, active, where } = req.body;
   try {
     const newSchedule = await ScheduleService.updateSchedule(ScheduleId, {
-      start,end, active, where
+      start,
+      end,
+      active,
+      where,
     });
     res.status(200).json(newSchedule);
   } catch (error) {
@@ -43,6 +68,10 @@ const updateSchedule = async (req, res) => {
 };
 
 const deleteSchedule = async (req, res) => {
+  const result = validationResult(req);
+  if (!result.isEmpty()) {
+     return res.status(403).send({ errors: result.array() });
+  }
   const ScheduleId = req.params.ScheduleId;
   try {
     const Schedule = await ScheduleService.deleteSchedule(ScheduleId);
